pyproject.toml
[tool.poetry]
name = "simple-internets"
version = "1.0.0"
description = "This is an internet speed test that will run on a client provided interval."
authors = ["PCDSandwichMan <46288251+PCDSandwichMan@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
speedtest-cli = "^2.1.3"
colorama = "^0.4.6"
alive-progress = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
